refactor(frontend): rename sections/ to chromes/, fold WikiHeader in
"sections" collides with the common convention of naming big blocks within
a single page (HeroSection, PricingSection, ...), which is the opposite of
what this directory holds: persistent app-shell chrome (sidebar, header)
mounted once by the layout and never torn down across navigations.
"chromes" is unambiguous and keeps the plural-noun convention already used
elsewhere (components, providers, views).

Also moves WikiHeader.tsx from components/wiki/ into chromes/header/ — it's
the same kind of thing as the sidebar (persistent shell, not a reusable
component, not a page), just previously stranded in the generic wiki
components bucket.
