+++ frontend/src/chromes/sidebar/AppSidebar.tsx
@@ -11,7 +11,7 @@ import {
   SvgTrash,
 } from "@onyx-ai/opal/icons";
 import { SidebarLayouts, useSidebarState } from "@onyx-ai/opal/layouts";
-import { sidebarLogo } from "@/sections/sidebar/shared";
+import { sidebarLogo } from "@/chromes/sidebar/shared";
 import { useRef } from "react";
 import useSWR from "swr";
 import {
@@ -23,9 +23,9 @@ import { useAuth } from "@/lib/auth";
 import { RECENTS_KEY, type RecentDocsResponse } from "@/lib/recents";
 import { STARRED_KEY, starDoc, type StarredDocsResponse } from "@/lib/starred";
 import { wikiHref } from "@/lib/wikiHref";
-import { docLabel } from "@/sections/sidebar/docLabel";
-import StarredList from "@/sections/sidebar/StarredList";
-import UserMenu from "@/sections/sidebar/UserMenu";
+import { docLabel } from "@/chromes/sidebar/docLabel";
+import StarredList from "@/chromes/sidebar/StarredList";
+import UserMenu from "@/chromes/sidebar/UserMenu";
 import { NAV_ENTRIES } from "@/lib/nav/registry";
 import { useIsMobile } from "@/lib/viewport";
 import { useAppFocus } from "@/hooks/useAppFocus";
