+++ frontend/src/chromes/sidebar/docLabel.ts
